news-aggregator/requirements.txt
flask
anthropic
feedparser
requests
beautifulsoup4
schedule
python-dotenv
praw
